Solve issue with sorted caused by premature optimization.
In rare cases, output of maps from `getSorted` would include data that wasn't supposed to be there.
In the name of optimization, we were trying to cache the sorted results via `updateCache` which had side effects due to expiry.

Instead, just provide `getSorted` and `getByResidency` and let clients decide :P